Improve file and text tools (#129)
* Improve tool descriptions and parameters and avoid changed in deleted files folders
* Support line numbers, and reading range of lines from file
* Support one replacement by default with option for multiple
* Add a tool to insert text at a line number
* Add a yielding method to make sure no work is done if not recording
* Improved line breaks in streaming mode

+++ src/tools/toolsets/text_editing/replace_text_file_tool.cr
@@ -7,17 +7,19 @@ module Tools::TextEditing
   # text in text-based files. It ensures the operation is performed securely within the
   # allowed directory, avoiding access to unauthorized paths.
   class ReplaceTextInTextFileTool < BuiltInFunction
-    name "replace_text_in_text_file"
+    name "str_replace_in_text_file"
 
-    description "Searches for the specified text in a text-based file and replaces it with alternative text. " +
-                "Ensures the file is within the current directory and is a text file."
+    description "Replaces a specified string in a text file within the current directory with a new string." \
+                "This is used for making precise edits."
 
     param "file_path", type: Param::Type::Str,
-      description: "The relative path to the file where you want to perform the replacement.", required: true
-    param "search", type: Param::Type::Str,
-      description: "The text to search for in the file.", required: true
-    param "replace_with", type: Param::Type::Str,
-      description: "The text to use to replace the text we searched for in the file.", required: true
+      description: "The relative path to the text file to modify.", required: true
+    param "old_str", type: Param::Type::Str,
+      description: "The text to replace (must match exactly, including whitespace and indentation)", required: true
+    param "new_str", type: Param::Type::Str,
+      description: "The new text to insert in place of the old text.", required: true
+    param "multiple", type: Param::Type::Bool, required: false,
+      description: "If true, tries to replaces multiple occurrences of the old string. Default is false, replacing only the first occurence."
 
     runner Runner
 
@@ -27,28 +29,37 @@ module Tools::TextEditing
 
       def execute(args : JSON::Any) : String
         file_path = args["file_path"]?.try &.as_s? ||
-                    return error_response("The required file_path was not specified")
-        search_text = args["search"]?.try &.as_s? ||
-                      return error_response("The required search_text was not specified")
-        replacement_text = args["replace_with"]?.try &.as_s? ||
-                           return error_response("The required replacement_text was not specified")
+                    return error_response("The required `file_path` was not specified")
+        search_text = args["old_str"]?.try &.as_s? ||
+                      return error_response("The required `old_str` was not specified")
+        replacement_text = args["new_str"]?.try &.as_s? ||
+                           return error_response("The required `new_str` was not specified")
+        multiple = args["multiple"]?.try &.as_bool? || false
 
         resolved_path = resolve_path(file_path)
 
         return error_response("The specified path '#{file_path}' is not allowed.") unless within_current_directory?(resolved_path)
-        return error_response("The specified file '#{file_path}' does not exist.") unless valid_file?(resolved_path)
+        return error_response("The specified file '#{file_path}' does not exist or is not a file.") unless valid_file?(resolved_path)
 
         begin
           changes = 0
           content = File.read(resolved_path)
-          new_content = content.gsub(search_text) do
-            changes += 1 # Count changes so we can detect if nothing changed
-            replacement_text
+          if multiple
+            new_content = content.gsub(search_text) do
+              changes += 1 # Count changes so we can detect if nothing changed
+              replacement_text
+            end
+          else
+            # Replace first occurrence only
+            new_content = content.sub(search_text) do
+              changes += 1 # Count changes so we can detect if nothing changed
+              replacement_text
+            end
           end
-          raise RuntimeError.new("Unable to find search text in the file") if changes.zero?
+          raise RuntimeError.new("Unable to find old string in the file. Nothing to replace.") if changes.zero?
           # Changes made
           File.write(resolved_path, new_content)
-          success_response(file_path, new_content)
+          success_response(file_path, new_content, changes)
         rescue e
           error_response("An error occurred while modifying the file: #{e.message}")
         end
@@ -58,8 +69,8 @@ module Tools::TextEditing
         {"error" => message}.to_json
       end
 
-      private def success_response(file_path, content)
-        {file_path: file_path, new_content: content}.to_json
+      private def success_response(file_path, content, changes : Int32)
+        {file_path: file_path, new_content: content, replacements: changes}.to_json
       end
     end
   end
